* Using `int_fast64_t` instead of `int32_t` - to accomodate larger values of `PRINTF_MAX_FLOAT`. TODO: We could theoretically adapt the type used to the value of `PRINTF_MAX_FLOAT`, using a macro.
* Added testcases which - hopefully - should hit both below and above the `PRINT_MAX_FLOAT_RANGE`.

+++ printf.c
@@ -382,7 +382,7 @@ static size_t _ftoa(out_fct_type out, char* buffer, size_t idx, size_t maxlen, d
     prec--;
   }
 
-  int whole = (int)value;
+  int_fast64_t whole = (int_fast64_t)value;
   double tmp = (value - whole) * pow10[prec];
   unsigned long frac = (unsigned long)tmp;
   diff = tmp - (double)frac;
